Fixing bug in defining expression
